ssh improvement

+++ build.d/05_ssh.sh
@@ -5,7 +5,9 @@
 [ -z "$BOOT_MNT" ] && error "boot mount unknown"
 [ -z "$PROJECT_DIR" ] && error "project dir unknown"
 
-touch "${BOOT_MNT}/ssh"
-
-AKF="${PROJECT_DIR}/authorized_keys"
-[ -f "$AKF" ] && cp "$AKF" "${BOOT_MNT}"
+if [ -f "${PROJECT_DIR}/ssh" ]
+then
+  touch "${BOOT_MNT}/ssh"
+  AKF="${PROJECT_DIR}/authorized_keys"
+  [ -f "$AKF" ] && cp "$AKF" "${BOOT_MNT}"
+fi
